pyproject.toml
[project]
name = "recipe-chatbot"
version = "0.1.0"
description = "Foundation project for the 2025 AI Evals Course"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.12",
  "httpx>=0.28.1",
  "litellm>=1.70.2",
  "mirascope>=1.24.0",
  "python-dotenv>=1.1.0",
  "rich>=14.0.0",
  "uvicorn>=0.34.2",
]
